requirements.txt
ShopifyAPI==12.0.0
pandas==2.0.3
python-dotenv==1.0.0
pathlib2==2.3.7
requests==2.31.0
